Ref: Move EchoReader to ui.EchoReaderScreen
